+++ app/src/main/java/com/ksfc/newfarmer/beans/GiftOrderListResult.java
@@ -98,7 +98,7 @@ public class GiftOrderListResult extends ResponseResult {
                 public String category;
                 public String thumbnail;
                 public int points;
-                public float marketPrice;
+                public double marketPrice;
                 public int online;
                 public int soldout;
             }
