+++ app/src/main/java/com/ksfc/newfarmer/beans/GoodsSum.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class GoodsSum implements Serializable {
 	public int state;
-	public float price;
-	public float discount;
+	public double price;
+	public double discount;
 	public int num;
 
-	public GoodsSum(int state, float price, float discount, int num) {
+	public GoodsSum(int state, double price, double discount, int num) {
 		super();
 		this.state = state;
 		this.price = price;
