+++ app/src/main/java/com/ksfc/newfarmer/beans/RscOrderDetailResult.java
@@ -158,7 +158,7 @@ public class RscOrderDetailResult extends ResponseResult {
             }
 
             public static class AdditionsEntity implements Serializable {
-                public float price;//附加选项价格
+                public double price;//附加选项价格
                 public String name;// 附加选项名称
                 public String ref;//附加选项_id
                 public String _id;//附加选项_id
